# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/trucks": {
      "post": {
        "operationId": "TrucksController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTruckDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Trucks"
        ]
      },
      "get": {
        "operationId": "TrucksController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Trucks"
        ]
      }
    },
    "/trucks/{id}": {
      "get": {
        "operationId": "TrucksController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Trucks"
        ]
      },
      "patch": {
        "operationId": "TrucksController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Trucks"
        ]
      },
      "delete": {
        "operationId": "TrucksController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Trucks"
        ]
      }
    },
    "/drivers": {
      "post": {
        "operationId": "DriversController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDriverDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Drivers"
        ]
      },
      "get": {
        "operationId": "DriversController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Drivers"
        ]
      }
    },
    "/drivers/{id}": {
      "get": {
        "operationId": "DriversController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Drivers"
        ]
      },
      "patch": {
        "operationId": "DriversController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Drivers"
        ]
      },
      "delete": {
        "operationId": "DriversController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Drivers"
        ]
      }
    },
    "/drivers/{id}/assign/{truckId}": {
      "patch": {
        "operationId": "DriversController_assignTruck",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "truckId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Drivers"
        ]
      }
    }
  },
  "info": {
    "title": "Transport Management API",
    "description": "Domentation for Transport Management API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateTruckDto": {
        "type": "object",
        "properties": {}
      },
      "CreateDriverDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}